Create folder structure pickers for picker pages and widgets/ for smaller widgets
