+++ lib/flutter_hsvcolor_picker.dart
@@ -1,14 +1,14 @@
 library flutter_hsvcolor_picker;
 
-export 'src/alpha_picker.dart';
+export 'src/widgets/alpha_picker.dart';
 export 'src/color_picker.dart';
-export 'src/hex_picker.dart';
-export 'src/hsv_picker.dart';
-export 'src/palette_hue_picker.dart';
-export 'src/palette_picker.dart';
-export 'src/palette_saturation_picker.dart';
-export 'src/palette_value_picker.dart';
-export 'src/rgb_picker.dart';
-export 'src/slider_picker.dart';
-export 'src/swatches_picker.dart';
-export 'src/wheel_picker.dart';
+export 'src/widgets/hex_picker.dart';
+export 'src/pickers/hsv_picker.dart';
+export 'src/pickers/palette_hue_picker.dart';
+export 'src/widgets/palette_picker.dart';
+export 'src/pickers/palette_saturation_picker.dart';
+export 'src/pickers/palette_value_picker.dart';
+export 'src/pickers/rgb_picker.dart';
+export 'src/widgets/slider_picker.dart';
+export 'src/pickers/swatches_picker.dart';
+export 'src/pickers/wheel_picker.dart';
