+++ lib/src/widgets/hex_picker.dart
