+++ lib/src/widgets/palette_picker.dart
