+++ lib/src/widgets/slider_picker.dart
